backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.0.3
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.12.0.88
setuptools==80.9.0
Werkzeug==3.1.3
wheel==0.45.1
pymysql==1.1.1
openai==1.12.0
Pillow==10.2.0
python-dotenv==1.0.0
